+++ src/Builder/IdentifierRenamer.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\NodeChanger;
+namespace Rector\Builder;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr\ClassConstFetch;
@@ -11,9 +11,6 @@ use PhpParser\Node\Identifier;
 use PhpParser\Node\Stmt\ClassMethod;
 use Rector\Exception\NodeChanger\NodeMissingIdentifierException;
 
-/**
- * @todo move to NodeBuilder
- */
 final class IdentifierRenamer
 {
     /**
