+++ src/Rector/Contrib/Nette/Forms/FormNegativeRulesRector.php
@@ -6,8 +6,8 @@ use PhpParser\Node;
 use PhpParser\Node\Expr\BitwiseNot;
 use PhpParser\Node\Expr\ClassConstFetch;
 use PhpParser\Node\Identifier;
+use Rector\Builder\IdentifierRenamer;
 use Rector\NodeAnalyzer\ClassConstAnalyzer;
-use Rector\NodeChanger\IdentifierRenamer;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
 use Rector\RectorDefinition\RectorDefinition;
