+++ src/Rector/Contrib/Symfony/Controller/ActionSuffixRemoverRector.php
@@ -4,7 +4,7 @@ namespace Rector\Rector\Contrib\Symfony\Controller;
 
 use PhpParser\Node;
 use Rector\Bridge\Symfony\NodeAnalyzer\ControllerMethodAnalyzer;
-use Rector\NodeChanger\IdentifierRenamer;
+use Rector\Builder\IdentifierRenamer;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
 use Rector\RectorDefinition\RectorDefinition;
