+++ src/Rector/Dynamic/ClassConstantReplacerRector.php
@@ -5,8 +5,8 @@ namespace Rector\Rector\Dynamic;
 use PhpParser\Node;
 use PhpParser\Node\Expr\ClassConstFetch;
 use PhpParser\Node\Identifier;
+use Rector\Builder\IdentifierRenamer;
 use Rector\NodeAnalyzer\ClassConstAnalyzer;
-use Rector\NodeChanger\IdentifierRenamer;
 use Rector\Rector\AbstractRector;
 use Rector\RectorDefinition\CodeSample;
 use Rector\RectorDefinition\RectorDefinition;
